+++ spec/html-inspector-spec.js
@@ -735,14 +735,14 @@ describe("validation", function() {
   })
 
   it("can determine if a child elememnt is allowed inside it's parent", function() {
-    expect(validation.isChildDisallowedInParent("div", "ul")).toBe(true)
-    expect(validation.isChildDisallowedInParent("div", "span")).toBe(true)
-    expect(validation.isChildDisallowedInParent("section", "em")).toBe(true)
-    expect(validation.isChildDisallowedInParent("title", "body")).toBe(true)
-    expect(validation.isChildDisallowedInParent("strong", "p")).toBe(false)
-    expect(validation.isChildDisallowedInParent("li", "ol")).toBe(false)
-    expect(validation.isChildDisallowedInParent("fieldset", "form")).toBe(false)
-    expect(validation.isChildDisallowedInParent("td", "tr")).toBe(false)
+    expect(validation.isChildAllowedInParent("div", "ul")).toBe(false)
+    expect(validation.isChildAllowedInParent("div", "span")).toBe(false)
+    expect(validation.isChildAllowedInParent("section", "em")).toBe(false)
+    expect(validation.isChildAllowedInParent("title", "body")).toBe(false)
+    expect(validation.isChildAllowedInParent("strong", "p")).toBe(true)
+    expect(validation.isChildAllowedInParent("li", "ol")).toBe(true)
+    expect(validation.isChildAllowedInParent("fieldset", "form")).toBe(true)
+    expect(validation.isChildAllowedInParent("td", "tr")).toBe(true)
   })
 
   it("ignores elements that are whitelisted", function() {
@@ -1104,76 +1104,6 @@ describe("inline-event-handlers", function() {
     })
 
     expect(log.length).toBe(0)
-  })
-
-})
-
-
-describe("scoped-styles", function() {
-
-  var log
-
-  function onComplete(reports) {
-    log = []
-    reports.forEach(function(report) {
-      log.push(report)
-    })
-  }
-
-  it("warns when style elements outside of the head do not declare the scoped attribute", function() {
-    var html = parseHTML(''
-          + '<section>'
-          + '  <style> .foo { } </style>'
-          + '</section>'
-        )
-
-    HTMLInspector.inspect({
-      useRules: ["scoped-styles"],
-      domRoot: html,
-      onComplete: onComplete
-    })
-
-    expect(log.length).toBe(1)
-    expect(log[0].message).toBe("<style> elements outside of <head> must declare the 'scoped' attribute.")
-    expect(log[0].context).toBe(html.querySelector("style"))
-
-  })
-
-  it("doesn't warns when style elements outside of the head declare the scoped attribute", function() {
-    var html = parseHTML(''
-          + '<section>'
-          + '  <style scoped> .foo { } </style>'
-          + '</section>'
-        )
-
-    HTMLInspector.inspect({
-      useRules: ["scoped-styles"],
-      domRoot: html,
-      onComplete: onComplete
-    })
-
-    expect(log.length).toBe(0)
-
-  })
-
-  it("doesn't warns when style elements are inside the head", function() {
-    var html = parseHTML(''
-          + '<html>'
-          + '  <head>'
-          + '    <style scoped> .foo { } </style>'
-          + '  </head>'
-          + '  <body></body>'
-          + '</html>'
-        )
-
-    HTMLInspector.inspect({
-      useRules: ["scoped-styles"],
-      domRoot: html,
-      onComplete: onComplete
-    })
-
-    expect(log.length).toBe(0)
-
   })
 
 })
@@ -1801,9 +1731,7 @@ describe("validate-element-location", function() {
     expect(log[5].context).toBe(html.querySelector("em > p"))
   })
 
-
   it("doesn't warn when elements appear as children of parents they're allowed to be within", function() {
-
     var html = parseHTML(''
           + '<div>'
           + '  <h1>This is a <strong>Heading!</strong> shit</h1>'
@@ -1814,15 +1742,76 @@ describe("validate-element-location", function() {
           + '  </section>'
           + '</div>'
         )
+    HTMLInspector.inspect({
+      useRules: ["validate-element-location"],
+      domRoot: html,
+      onComplete: onComplete
+    })
+    expect(log.length).toBe(0)
+  })
+
+  it("warns when <style> elements inside body do not declare the scoped attribute", function() {
+    var html = document.createElement("body")
+    html.innerHTML = '<section><style> .foo { } </style></section>'
 
     HTMLInspector.inspect({
       useRules: ["validate-element-location"],
       domRoot: html,
       onComplete: onComplete
     })
+    expect(log.length).toBe(1)
+    expect(log[0].message).toBe("<style> elements inside <body> must contain the 'scoped' attribute.")
+    expect(log[0].context).toBe(html.querySelector("style"))
+  })
 
+  it("doesn't warns when <style> elements are inside the head", function() {
+    var html = parseHTML(''
+          + '<html>'
+          + '  <head>'
+          + '    <style scoped> .foo { } </style>'
+          + '  </head>'
+          + '  <body></body>'
+          + '</html>'
+        )
+    HTMLInspector.inspect({
+      useRules: ["scoped-styles"],
+      domRoot: html,
+      onComplete: onComplete
+    })
     expect(log.length).toBe(0)
+  })
 
+  it("warns when <link> and <meta> elements inside body do not declare the itemprop attribute", function() {
+    var html = document.createElement("body")
+    html.innerHTML = '<meta charset="utf-8"><link rel="imports" href="component.html">'
+    HTMLInspector.inspect({
+      useRules: ["validate-element-location"],
+      domRoot: html,
+      onComplete: onComplete
+    })
+    expect(log.length).toBe(2)
+    expect(log[0].message).toBe("<meta> elements inside <body> must contain the 'itemprop' attribute.")
+    expect(log[0].context).toBe(html.querySelector("meta"))
+    expect(log[1].message).toBe("<link> elements inside <body> must contain the 'itemprop' attribute.")
+    expect(log[1].context).toBe(html.querySelector("link"))
+  })
+
+  it("doesn't warns when <link> and <meta> elements are inside the head", function() {
+    var html = parseHTML(''
+          + '<html>'
+          + '  <head>'
+          + '    <meta charset="utf-8">'
+          + '    <link rel="imports" href="component.html">'
+          + '  </head>'
+          + '  <body></body>'
+          + '</html>'
+        )
+    HTMLInspector.inspect({
+      useRules: ["validate-element-location"],
+      domRoot: html,
+      onComplete: onComplete
+    })
+    expect(log.length).toBe(0)
   })
 
 })